Race mapping

+++ src/main/java/com/room414/racingbets/dal/domain/infrastructure/BuildHelper.java
@@ -16,7 +16,7 @@ public class BuildHelper {
 
     }
 
-    public static List<Participant> mapIdsToParticipants(List<Integer> ids) {
+    public static List<Participant> mapIdsToParticipants(List<Long> ids) {
         if (ids != null) {
             return ids.stream().map((id) -> {
                 Participant participant = new Participant();
